backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
redis>=4.6.0
celery>=5.3.0
langchain>=0.0.260
dashscope>=1.14.0
httpx>=0.24.0
motor>=3.2.0
pandas>=2.0.0
openpyxl>=3.1.0
oss2>=2.18.4
greenlet>=3.0.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

